refactor: move helper functions that are dependent on shelljs to cli package
