+++ packages/cli/src/sde-exec.js
@@ -1,5 +1,6 @@
 import path from 'path'
-import { modelPathProps, buildDir, outputDir, execCmd } from './Helpers.js'
+import { modelPathProps, buildDir, outputDir } from './Helpers.js'
+import { execCmd } from './utils.js'
 
 export let command = 'exec [options] <model>'
 export let describe = 'execute the model and capture its output to a file'
