--- packages/compile/src/index.js
@@ -1,1 +1,0 @@
-// TODO
